misc bug fixes

+++ auth/lib/jdomodels/src/main/java/org/sagebionetworks/repo/model/jdo/JDODatasetDAOImpl.java
@@ -5,9 +5,11 @@ import java.util.Collection;
 import java.util.HashSet;
 import java.util.Set;
 
+import javax.jdo.JDOObjectNotFoundException;
 import javax.jdo.PersistenceManager;
 import javax.jdo.Transaction;
 
+import org.sagebionetworks.repo.model.AuthorizationConstants;
 import org.sagebionetworks.repo.model.Dataset;
 import org.sagebionetworks.repo.model.DatasetDAO;
 import org.sagebionetworks.repo.model.DatastoreException;
@@ -159,14 +161,27 @@ public class JDODatasetDAOImpl extends
 	public void delete(String id) throws DatastoreException, NotFoundException,
 	UnauthorizedException {
 		PersistenceManager pm = PMF.get();
-//		Transaction tx = pm.currentTransaction();
-		JDODataset ds = pm.getObjectById(JDODataset.class, KeyFactory.stringToKey(id));
+	try {
+		Long key = KeyFactory.stringToKey(id);
+		JDODataset ds = pm.getObjectById(JDODataset.class, key);
+		//  authorization check comes AFTER the retrieval step, so that we get a 'not found' result
+		// rather than 'forbidden' when an object does not exist.
+		if (!JDOUserGroupDAOImpl.canAccess(userId, getJdoClass().getName(), key, AuthorizationConstants.CHANGE_ACCESS, pm))
+			throw new UnauthorizedException();
 		JDOInputDataLayerDAOImpl idlDAO = (JDOInputDataLayerDAOImpl)getInputDataLayerDAO(id);
 		Set<JDOInputDataLayer> idls = ds.getInputLayers();
 		for (JDOInputDataLayer idl: idls) {
 			idlDAO.removeResourceFromAllGroups(idl.getId());
 		}
-//		tx.commit();
+	} catch (JDOObjectNotFoundException e) {
+		throw new NotFoundException(e);
+	} catch (UnauthorizedException e) {
+		throw e;
+	} catch (Exception e) {
+		throw new DatastoreException(e);
+	} finally {
+		pm.close();
+	}
 		super.delete(id);
 	}
